==> 02_saving_data_into_db.ipynb <==
# Save Bus Data into Database
In this notebook, we take the cleaned bus data from CSV files and save it into a database.  

- We use **SQLite** because it is simple, file-based, and does not require a separate database server.  
- We also demonstrate saving data to **PostgreSQL**, which is a powerful relational database suitable for larger projects and multi-user access.


import pandas as pd
import sqlite3
import os
from sqlalchemy import create_engine
from urllib.parse import quote_plus


## Load CSV File


csv_path = "Data/All_Routes_Buses_Final.csv"

if not os.path.exists(csv_path):
    raise FileNotFoundError("CSV file not found. Run 01_web_scraping.ipynb first.")


**Read CSV into DataFrame**

Load the cleaned bus data from the CSV file into a Pandas DataFrame.  


df = pd.read_csv(csv_path)
df.head()


## Create SQLite Database

- Define the path for the SQLite database file (`bus_data.db`) inside the `data` folder.  
- Connect to the database. If the file does not exist, SQLite will create it automatically.


db_path = "data/bus_data.db"
conn = sqlite3.connect(db_path)


**Save Data to SQLite and Verify**

Save the `DataFrame` into the SQLite database as a table named `buses`.  



df.to_sql("buses", conn, if_exists="replace", index=False)
df_check = pd.read_sql("SELECT * FROM buses LIMIT 5", conn)
print("Data saved to SQLite database successfully")

print(df_check)
conn.close()



## Save Data to PostgreSQL
 
- Set up PostgreSQL connection details.  
- Encode the password to handle special characters and create a SQLAlchemy engine.  
- Save the DataFrame into PostgreSQL as a table named `bus_routes` 


# Load your CSV file
df = pd.read_csv(csv_path)


# PostgreSQL connection details
db_user = "postgres"
db_password = "Ganesha123@!"
db_host = "localhost"
db_port = "5432"
db_name = "bus_data"  # your database in pgAdmin

# Encode password to handle special characters
db_password_encoded = quote_plus(db_password)

# Create SQLAlchemy engine
engine = create_engine(f"postgresql+psycopg2://{db_user}:{db_password_encoded}@{db_host}:{db_port}/{db_name}")

# Store DataFrame in PostgreSQL
df.to_sql("bus_routes", engine, if_exists="replace", index=False)

# Verify by reading first 5 rows
df_check = pd.read_sql("SELECT * FROM bus_routes LIMIT 5", engine)
print("Data saved to PostgreSQL database successfully")
print(df_check)
